move antenna specific code to commnet page
