+++ src/app/overlays/celestial-body-details-dialog/planetoid-details.ts
@@ -1,4 +1,4 @@
-import { Antenna } from '../../common/domain/antenna';
+import { Antenna } from '../../pages/commnet-planner/models/antenna';
 import { PlanetoidType } from '../../common/domain/space-objects/planetoid-type';
 
 export class PlanetoidDetails {
