+++ src/main.ts
@@ -1,7 +1,10 @@
 import { enableProdMode } from '@angular/core';
 import { bootstrapApplication } from '@angular/platform-browser';
+import {
+  appOptions,
+  appRoot,
+} from './app/app.bootstrap';
 import { environment } from './environments/environment';
-import { appOptions, appRoot } from './app/app.bootstrap';
 
 import './extensions';
